% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondisc_preprocessing.R
\name{check_ondisc_inputs}
\alias{check_ondisc_inputs}
\title{Check ondisc inputs}
\usage{
check_ondisc_inputs(
  mm_odm,
  response_grna_group_pairs,
  form,
  response_modality_name,
  grna_modality_name,
  grna_group_column_name,
  low_moi
)
}
\arguments{
\item{mm_odm}{a multimodal ondisc matrix containing response and grna (either expression or assignment) modalities}

\item{response_grna_group_pairs}{a data frame with columns \code{response_id} and \code{grna_group} giving the response ID / grna group pairs to analyze.}

\item{form}{a formula object (or string) specifying how to adjust for the covariates}

\item{response_modality_name}{name of the response modality (e.g., "gene") within the multimodal ODM}

\item{grna_modality_name}{name of the grna modality (e.g., "grna_expression") within the multimodal ODM}

\item{grna_group_column_name}{name of the column within the feature covariate matrix of the gRNA ODM that contains the gRNA groups}

\item{low_moi}{does this dataset correspond to low MOI (TRUE) or high MOI (FALSE)?}
}
\value{
a thinned and verified multimodal ondisc matrix
}
\description{
Checks inputs (that come in ondisc format)
}
