% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_distributions.R
\name{compute_empirical_p_value}
\alias{compute_empirical_p_value}
\title{Compute empirical p-value}
\usage{
compute_empirical_p_value(z_star, z_null, side)
}
\arguments{
\item{z_star}{ground truth test statistic}

\item{z_null}{the null test statistics}

\item{side}{side of the tests}
}
\value{
the empirical p-value
}
\description{
Computes an empirical permutation test p-value.
}
