% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{compute_empirical_p_value_from_batch_result}
\alias{compute_empirical_p_value_from_batch_result}
\title{Compute empirical p-value from gRNA-wise resul}
\usage{
compute_empirical_p_value_from_batch_result(grna_wise_result, B, side)
}
\arguments{
\item{grna_wise_result}{data table giving the number of test statistics falling to the left of the original test statistic across batches (with genes in columns and rows in batches)}

\item{B}{total number of resamples}

\item{side}{sidedness of the test}
}
\value{
a vector of gene-wise p-values
}
\description{
Compute empirical p-value from gRNA-wise resul
}
