% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table.R
\name{compute_fisher_exact_p_value}
\alias{compute_fisher_exact_p_value}
\title{Compute Fisher's exact p-value}
\usage{
compute_fisher_exact_p_value(
  n_treatment_cells_with_expression,
  n_treatment_cells_without_expression,
  n_control_cells_with_expression,
  n_control_cells_without_expression
)
}
\arguments{
\item{treatment_expressions}{the vector of expressions for the treatment cells (i.e., the cells receiving a targeting gRNA)}

\item{control_expressions}{the vector of expressions for the control cells (i.e., the cells receiving an NT)}
}
\value{
p-value from the Fisher exact test for two-way contingency tables
}
\description{
Compute Fisher's exact p-value
}
