% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_distributions.R
\name{compute_ks_test}
\alias{compute_ks_test}
\title{Compute KS test}
\usage{
compute_ks_test(z_null, dp, distribution)
}
\arguments{
\item{z_null}{the vector of null test statistics}

\item{dp}{the fitted parameters of the distribution}

\item{distribution}{the distribution, either "SN" (for skew-normal) or "ST" (for skew-t)}
}
\value{
a list containing (i) the KS test statistic, and (ii) the KS p-value
}
\description{
Runs a KS test on the resampled test statistics
}
