% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_association_testing.R
\name{compute_nb_test_stat}
\alias{compute_nb_test_stat}
\title{Compute NB test statistic}
\usage{
compute_nb_test_stat(y, mu, response_theta)
}
\arguments{
\item{y}{vector of expressions}

\item{mu}{vector of means}

\item{response_theta}{estimated theta of response expression distribution}
}
\value{
the test statistic
}
\description{
Compute NB test statistic
}
