% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_t.R
\name{fit_skew_t}
\alias{fit_skew_t}
\title{Fit skew t}
\usage{
fit_skew_t(y, max_it = 70)
}
\arguments{
\item{y}{a numeric vector of data.}

\item{max_it}{maximum number of iterations.}
}
\value{
a list containing (i) dp, the fitted parameters and (ii) convergence success (TRUE or FALSE)
}
\description{
Fits a skew-t distribution to a vector of data y.
}
\examples{
y <- sn::rst(n = 2500, dp = c(0, 0.5, 2, 20))
fit <- fit_skew_t(y)
plot_skew_t(fit$dp, y)

y <- rpois(2000, 0.01)
fit <- fit_skew_t(y)

sn::selm(y ~ 1, family = "ST")@param$dp
}
