% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{get_grna_group_info}
\alias{get_grna_group_info}
\title{Get gRNA index info}
\usage{
get_grna_group_info(grna_group_assignments)
}
\arguments{
\item{grna_group_assignments}{a vector of cell-to-gRNA assignments}
}
\value{
a list containing (i) the indexes of the cells to which each gRNA belongs, and (ii) the number of cells in which each gRNA belongs
}
\description{
Returns relevant information about the positions of the gRNAs among the cells.
}
