% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{get_grna_permutation_idxs}
\alias{get_grna_permutation_idxs}
\title{Get gRNA permutation indices}
\usage{
get_grna_permutation_idxs(n_cells_per_grna, unique_grna, B)
}
\arguments{
\item{n_cells_per_grna}{the "n_cells_per_grna" table of the grna_group_info list}

\item{unique_grna}{a gRNA}

\item{B}{the number of resamples to draw}
}
\value{
a matrix of permutation indices
}
\description{
Obtains the matrix of permutation indices for a given gRNA and choice of B
}
