% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_precomputation.R
\name{get_pieces_from_gene_precomp}
\alias{get_pieces_from_gene_precomp}
\title{Get pieces from gene precomp}
\usage{
get_pieces_from_gene_precomp(gene_precomp, global_cell_covariates)
}
\arguments{
\item{gene_precomp}{a gene precomp vector}

\item{global_cell_covariates}{a global cell covariates matrix}
}
\value{
a list containing (i) gene_theta, and (ii) the fitted linear components
}
\description{
Get pieces from gene precomp
}
