% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_precomputation.R
\name{get_pieces_from_response_precomp}
\alias{get_pieces_from_response_precomp}
\title{Get pieces from response precomp}
\usage{
get_pieces_from_response_precomp(response_precomp, global_cell_covariates)
}
\arguments{
\item{response_precomp}{a response precomp vector}

\item{global_cell_covariates}{a global cell covariates matrix}
}
\value{
a list containing (i) response_theta, and (ii) the fitted linear components
}
\description{
Get pieces from response precomp
}
