% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_distributions.R
\name{plot_fitted_density}
\alias{plot_fitted_density}
\title{Plot fitted density}
\usage{
plot_fitted_density(dp, z_null, distribution, z_star = NULL, legend = TRUE)
}
\arguments{
\item{dp}{vector of fitted skew-normal or skew-t parameters}

\item{z_null}{vector of resampled test statistics to which the model was fitted}

\item{distribution}{either "ST" (for skew-t) or "SN" (for skew-normal)}

\item{z_star}{(optional) ground truth test statistic}
}
\value{
a ggplot object
}
\description{
Plots a fitted skew-normal or skew-t density superimposed over the histogram of resampled test statistics to which the density was fitted.
}
