% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_t.R
\name{plot_skew_t}
\alias{plot_skew_t}
\title{Plot skew-t}
\usage{
plot_skew_t(dp, z_null, z_star = NULL)
}
\arguments{
\item{dp}{vector of fitted skew-t parameters}

\item{z_null}{vector of resampled test statistics to which skew-t was fitted}

\item{z_star}{(optional) ground truth test statistic}
}
\value{
a ggplot object
}
\description{
Plot skew-t
}
