% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_precomputation.R
\name{run_grna_precomputation_low_level}
\alias{run_grna_precomputation_low_level}
\title{Run grna precomputation}
\usage{
run_grna_precomputation_low_level(indicators, covariate_matrix)
}
\arguments{
\item{indicators}{the binary vector of grna indicators}

\item{covariate_matrix}{the covariate matrix on which to regress (NOTE: should contain an interecept term)}
}
\value{
a named vector of fitted coefficients
}
\description{
Runs precomputation on a vector of grna indicators.
This function is used for conditional resampling tests only.
}
