% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_association_testing.R
\name{run_permutation_test}
\alias{run_permutation_test}
\title{Run permutation test}
\usage{
run_permutation_test(
  expressions,
  fitted_means,
  ground_truth_treatment_idxs,
  synthetic_treatment_idxs,
  response_theta,
  side
)
}
\arguments{
\item{expressions}{a vector of response expressions}

\item{fitted_means}{the vector of fitted means, obtained from regressing the expressions onto the technical factors}

\item{ground_truth_treatment_idxs}{indexes of the "treatment" cells on the ground truth data}

\item{synthetic_treatment_idxs}{indexes of the "treatment" cells under the resampled data}

\item{response_theta}{estimated theta of response expression distribution}

\item{side}{sidedness of the test (one of "left", "right", or "both")}

\item{full_output}{return the full output (TRUE) or reduced output (FALSE)?}
}
\description{
Run permutation test
}
