% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_association_testing.R
\name{run_permutations}
\alias{run_permutations}
\title{Run permutations}
\usage{
run_permutations(
  expressions,
  fitted_means,
  ground_truth_treatment_idxs,
  synthetic_treatment_idxs,
  gene_theta
)
}
\arguments{
\item{expressions}{a vector of gene expressions}

\item{fitted_means}{the vector of fitted means, obtained from regressing the expressions onto the technical factors}

\item{ground_truth_treatment_idxs}{indexes of the "treatment" cells on the ground truth data}

\item{synthetic_treatment_idxs}{indexes of the "treatment" cells under the resampled data}

\item{gene_theta}{estimated theta of gene expression distribution}
}
\value{
a list containing (i) z_star (the ground truth test statistic), (ii) z_null (the vector of resampled test statistics), and (iii) log_fold_change (the estimated log fold change)
}
\description{
Runs permutations using the NB score-test-based z statistic.
}
