% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_distributions.R
\name{fit_skew_t}
\alias{fit_skew_t}
\alias{fit_skew_normal}
\title{Fit a skew distribution}
\usage{
fit_skew_t(y, max_it = 70)

fit_skew_normal(y, max_it = 70)
}
\arguments{
\item{y}{a numeric vector of data.}

\item{max_it}{maximum number of iterations.}
}
\value{
a list containing (i) dp (the fitted model parameters), (ii) convergence success (TRUE or FALSE), and (iii) the number of iterations
}
\description{
Fit a skew-t or skew-normal distribution on a vector of data.
}
\examples{
y <- sn::rsn(n = 2500, dp = c(-.5, 1.5, 2))
fit_st <- fit_skew_t(y)
fit_sn <- fit_skew_normal(y)
plot_fitted_density(fit_st$dp, y, "ST")
plot_fitted_density(fit_sn$dp, y, "SN")

y <- rpois(2000, 0.01)
fit_st <- fit_skew_t(y)
fit_sn <- fit_skew_normal(y)
}
