% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level_precomputation.R
\name{run_response_precomputation_low_level}
\alias{run_response_precomputation_low_level}
\title{Run response precomputation}
\usage{
run_response_precomputation_low_level(expressions, covariate_matrix)
}
\arguments{
\item{expressions}{the numeric vector of response expressions}

\item{covariate_matrix}{the covariate matrix on which to regress (NOTE: should contain an interecept term)}
}
\value{
a named vector of fitted coefficients, alongside the fitted size parameter (named "response_theta")
}
\description{
Runs precomputation on a response
}
